[TNT-309] refactor: 기존 TextField 제거 및 "2" 접미사 제거 처리
